simulation/scripts/requirements.txt
cryptography==38.0.1
pandas==1.5.0
matplotlib==3.6.0
seaborn==0.12.0
PyYAML==6.0
jproperties==2.1.1
tikzplotlib==0.10.1